requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
pickle-mixin
torch
